Quora questions flow implementation

+++ quora-service/src/main/java/com/upgrad/quora/service/business/QuestionService.java
@@ -29,38 +29,78 @@ public class QuestionService {
 
     @Value("${user.default.role}")
     private String role;
-
+    /*
+    * @param questionEntity - questionEntity Object to be persisted
+    * @param authorizationToken - authorizationToken to be checked for authentication and authorization
+    * @throws AuthorizationFailedException - thrown if user does not exist or is signed out
+    * @returns questionEntity - questionEntity Object which is persisted
+    */
     @Transactional(propagation = Propagation.REQUIRED)
     public QuestionEntity createQuestion(final QuestionEntity questionEntity,final String authorizationToken) throws AuthorizationFailedException{
         UserAuthTokenEntity userAuthTokenEntity = checkUserAuth(authorizationToken);
+        if(userAuthTokenEntity.getLogoutAt()!=null){
+            throw new AuthorizationFailedException(ATHR_QSN_USR_001_COMMON.getCode(), ATHR_QSN_USR_001_COMMON.getDefaultMessage());
+        }
         questionEntity.setUser(userAuthTokenEntity.getUser());
         return questionDao.createQuestion(questionEntity);
     }
-
+    /*
+    * @param authorizationToken - authorizationToken to be checked for authentication and authorization
+    * @returns List<QuestionEntity> - returns list of question entities
+    */
     public List<QuestionEntity> getAllQuestions(final String authorizationToken) throws AuthorizationFailedException{
-        checkUserAuth(authorizationToken);
+        UserAuthTokenEntity userAuthTokenEntity=checkUserAuth(authorizationToken);
+        if(userAuthTokenEntity.getLogoutAt()!=null){
+            throw new AuthorizationFailedException(ATHR_QSN_USR_002_COMMON.getCode(), ATHR_QSN_USR_002_COMMON.getDefaultMessage());
+        }
         return questionDao.getQuestions();
     }
-
+    /*
+    * @param userId - user Id for retrieving user questions
+    * @param authorizationToken - authorizationToken to be checked for authentication and authorization
+    * @returns List<QuestionEntity> - returns list of question entities
+    */
     public List<QuestionEntity> getQuestionsByUser(final String userId,final String authorizationToken) throws AuthorizationFailedException,UserNotFoundException{
-        checkUserAuth(authorizationToken);
+        UserAuthTokenEntity userAuthTokenEntity=checkUserAuth(authorizationToken);
+        if(userAuthTokenEntity.getLogoutAt()!=null){
+            throw new AuthorizationFailedException(ATHR_QSN_USR_003_COMMON.getCode(), ATHR_QSN_USR_003_COMMON.getDefaultMessage());
+        }
         if(userDao.getUser(userId)==null){
             throw new UserNotFoundException(QSN_USER_001.getCode(), QSN_USER_001.getDefaultMessage());
         }
         return questionDao.getQuestionsByUser(userId);
     }
+
+    /*
+     * @param questionId - questionId Id for retrieving question to be deleted
+     * @param authorizationToken - authorizationToken to be checked for authentication and authorization
+     * @returns QuestionEntity - returns questionEntity which is deleted
+     */
     @Transactional(propagation = Propagation.REQUIRED)
     public QuestionEntity deleteQuestionById(final String questionId,final String authorizationToken) throws AuthorizationFailedException, InvalidQuestionException{
         UserAuthTokenEntity userAuthTokenEntity = checkUserAuth(authorizationToken);
+        if(userAuthTokenEntity.getLogoutAt()!=null){
+            throw new AuthorizationFailedException(ATHR_QSN_USR_004_COMMON.getCode(), ATHR_QSN_USR_004_COMMON.getDefaultMessage());
+        }
         QuestionEntity questionEntity = getQuestionByQuestionId(questionId);
         if(questionEntity.getUser()!=userAuthTokenEntity.getUser() && userAuthTokenEntity.getUser().getRole().equals(role)){
             throw new AuthorizationFailedException(ATHR_QSN_001_COMMON.getCode(),ATHR_QSN_001_COMMON.getDefaultMessage());
         }
         return questionDao.deleteQuestion(questionEntity);
     }
+
+    /*
+     * @param questionEntity - question Entity object
+     * @param questionId - questionId Id for retrieving question to be deleted
+     * @param authorizationToken - authorizationToken to be checked for authentication and authorization
+     * @returns QuestionEntity - returns questionEntity which is deleted
+     */
     @Transactional(propagation = Propagation.REQUIRED)
     public QuestionEntity editQuestion(final QuestionEntity questionEntity,final String questionId,final String authorizationToken)throws AuthorizationFailedException, InvalidQuestionException{
         UserAuthTokenEntity userAuthTokenEntity = checkUserAuth(authorizationToken);
+        if(userAuthTokenEntity.getLogoutAt()!=null){
+            throw new AuthorizationFailedException(ATHR_QSN_USR_005_COMMON.getCode(), ATHR_QSN_USR_005_COMMON.getDefaultMessage());
+        }
         QuestionEntity question = getQuestionByQuestionId(questionId);
         if(question.getUser()!=userAuthTokenEntity.getUser()){
             throw new AuthorizationFailedException(ATHR_QSN_002_COMMON.getCode(),ATHR_QSN_002_COMMON.getDefaultMessage());
@@ -69,17 +109,23 @@ public class QuestionService {
         question.setDate(questionEntity.getDate());
         return questionDao.editQuestion(question);
     }
-
+    /*
+    * checks user authentication based on the authorizationToken
+    * @param authorizationToken - authorizationToken to be checked for authentication and authorization
+    * @throws AuthorizationFailedException - thrown if user does not exist or is signed out
+    * @returns userAuthToken - UserAuthTokenEntity object for the authorizationToken
+    */
     private UserAuthTokenEntity checkUserAuth(final String authorizationToken) throws AuthorizationFailedException{
         UserAuthTokenEntity userAuthTokenEntity= userDao.getUserAuthToken(authorizationToken);
         if(userAuthTokenEntity==null) {
             throw new AuthorizationFailedException(ATHR_001_COMMON.getCode(), ATHR_001_COMMON.getDefaultMessage());
-        }else if(userAuthTokenEntity.getLogoutAt()!=null){
-            throw new AuthorizationFailedException(ATHR_002_COMMON.getCode(), ATHR_002_COMMON.getDefaultMessage());
         }
         return userAuthTokenEntity;
     }
-
+    /*
+    * @param questionId
+    * @returns questionEntity
+     */
     private QuestionEntity getQuestionByQuestionId(final String questionId) throws InvalidQuestionException{
         QuestionEntity question = questionDao.getQuestionById(questionId);
         if(question==null){
